Iteration type fixes
Various index types were fixed to match the loop test value, mostly
converting to `unsigned long` to accommodate the potentially very long
loop counts to reach O(1 sec) loop times.

+++ src/main.c
@@ -159,7 +159,7 @@ int main(int argc, char *argv[])
     timer = stopwatch_create(cfg->timer_type);
     do {
         timer->start(timer);
-        for (int i = 0; i < iter; i++)
+        for (unsigned long i = 0; i < iter; i++)
             a++;
         timer->stop(timer);
         iter *= 2;
